Allow time zone specification for expiresAt values.
This change introduces a 'timezone' option for cache policy objects
that determines which time zone the 'expiresAt' time is relative to.
Time zones may be specified either as a fixed offset (ex. -0800) or
as a IANA time zone name (ex. America/Los_Angeles). When using a
named time zone, daylight savings time will be accounted for
automatically.

Closes #112.